fix websocket issue

+++ start.js
@@ -3,6 +3,10 @@ import {spawn} from 'node:child_process'
 import fs from 'fs-extra'
 import {fileURLToPath} from 'url'
 import path from 'path'
+import yargs from 'yargs/yargs'
+import {hideBin} from 'yargs/helpers'
+const argv = yargs(hideBin(process.argv)).argv
+console.log({argv})
 const rootPath = path.dirname(fileURLToPath(import.meta.url))
 const reportPath = path.join(rootPath, 'report.json')
 import benchmarkOptionsFile from './benchmark-options.js'
@@ -34,15 +38,21 @@ fs.removeSync(reportPath)
 
 console.log(resolveAddressesBenchmarkOptions)
 
+const isOnly = (name) => argv.only === name || (argv.only?.length || 0) < 1 || argv.only.includes(name)
+
 // benchmark resolve addresses
-for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
-  fs.removeSync(benchmarkOptions.plebbitOptions.dataPath)
-  const benchmarkFile = 'benchmark-resolve-addresses.js'
-  await benchmarkNode(benchmarkFile, benchmarkOptions)
+if (isOnly('node')) {
+  for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
+    fs.removeSync(benchmarkOptions.plebbitOptions.dataPath)
+    const benchmarkFile = 'benchmark-resolve-addresses.js'
+    await benchmarkNode(benchmarkFile, benchmarkOptions)
+  }
 }
-for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
-  const benchmarkFile = 'benchmark-resolve-addresses.js'
-  await benchmarkChrome(benchmarkFile, benchmarkOptions)
+if (isOnly('chrome')) {
+  for (const benchmarkOptions of resolveAddressesBenchmarkOptions) {
+    const benchmarkFile = 'benchmark-resolve-addresses.js'
+    await benchmarkChrome(benchmarkFile, benchmarkOptions)
+  }
 }
 
 await printReport()
